% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GORICA-benchmarks_restriktor_ANOVA.R
\name{benchmarks_ANOVA}
\alias{benchmarks_ANOVA}
\title{Benchmarks for GORIC(A) weights}
\usage{
benchmarks_ANOVA(
  goric_obj,
  pop.es = 0.2,
  ratio.pop.means = NULL,
  other.N = NULL,
  CI.iter = 1000,
  seed.value = 123
)
}
\arguments{
\item{goric_obj}{An object from the goric function from the restriktor package. In this function, it is assumed that the GORIC is applied to an ANOVA with k group means.}

\item{pop.es}{Optional. A scalar or vector of population Cohen's d (effect size) value. By default, pop.es = .2. Benchmarks will be calculated for each of these value(s).}

\item{ratio.pop.means}{Optional. A k-times-1 vector, denoting the relative difference between the k group means. Note that a ratio of c(3,2,1) gives the same as c(1,0,-1), since the consecutive relative differences are 1 in both ratios. every time minus same difference. By default, ratio.pop.means = NULL. In that case, the relative differences from the data are used.}

\item{other.N}{Optional. A k-times-1 vector or a scalar to denote the group sizes, if you like to use others than used in the data. You could use this, for instance, to see to which values the GORIC(A) weights will converge (and thus to see the maximum value of the weights). If you specify a scalar, it is assumed that each group is of that size. By default, other.N = NULL. In that case, the group sizes from the data will be used.}

\item{CI.iter}{Optional. A scalar denoting the number of iterations used to determine the benchmarks weights and their 95\\% confidence intervals (CIs). By default, CI.iter = 1000. Notable, the higher CI.iter, the longer the computation time.}

\item{seed.value.}{Optional. A scalar denoting the seed value. By default, seed.value = 123. By changing this value, you can inspect the sensitivity of the benchmarks -- which should not be present; if it is, you should increase CI.iter.}
}
\value{
The output comprises case-specific benchmarks for GORIC(A) weights (for each of the specified effect sizes) and their 95\\% confidence intervals (CIs).
}
\description{
This function calculates case-specific benchmarks for GORIC(A) weights, together with their 95\\% confidence intervals (CIs).
}
\examples{

# library(benchmarks)

# Input 'data'
goric.obj <- My_goric_obj

# Input w.r.t. population values
pop.es <- c(0, .2, .5, .8)
ratio.pop.means <- c(3,2,1)

# Extra
iter <- 100

# Calculate case-specific benchmarks and their CIs
benchmarks_goric <- benchmarks_ANOVA(goric.obj, pop.es, ratio.pop.means, CI.iter = iter)
benchmarks_goric$pop.means
benchmarks_goric$benchmarks
benchmarks_goric$CI.benchmarks

# An example to see what maximum value of the weights is.
# If there is a maximum, then there is overlap in the hypotheses (see guidelines).
pop.es <- c(.2, .5, .8)
benchmarks_goric_1000 <- benchmarks_ANOVA(goric.obj, pop.es, ratio.pop.means, other.N = 1000, CI.iter = iter)
benchmarks_goric_1000$benchmarks
benchmarks_goric_1000$CI.benchmarks

}
